# Rewrite min and max hybrid handlers