--- a/inst/include/dplyr/Result/MinMax.h
+++ b/inst/include/dplyr/Result/MinMax.h
@@ -16,59 +16,6 @@ namespace dplyr {
   private:
     static const STORAGE Inf;
 
-    class ChunkProcessor {
-    public:
-      ChunkProcessor(const STORAGE* data_ptr_, const SlicingIndex& indices_) :
-        data_ptr(data_ptr_), indices(indices_), i(0), n(indices_.size()), res(Inf) {}
-
-      STORAGE process() {
-        process_loop<true>();
-        process_loop<false>();
-        return res;
-      }
-
-    private:
-      template <bool ONLY_CHECK_NA>
-      void process_loop() {
-        for (; i < n; ++i) {
-          STORAGE current = data_ptr[indices[i]];
-
-          if (Rcpp::Vector<RTYPE>::is_na(current)) {
-            if (NA_RM)
-              continue;
-            else {
-              res = current;
-              return;
-            }
-          }
-
-          if (ONLY_CHECK_NA) {
-            res = current;
-            ++i;
-            return;
-          }
-          else {
-            if (is_better(current, res)) res = current;
-          }
-        }
-      }
-
-      inline bool is_better(const STORAGE& current, const STORAGE& res) {
-        if (MINIMUM)
-          return internal::is_smaller<RTYPE>(current, res);
-        else
-          return internal::is_smaller<RTYPE>(res, current);
-      }
-
-    private:
-      const STORAGE* data_ptr;
-      const SlicingIndex& indices;
-
-      int i;
-      int n;
-      STORAGE res;
-    };
-
   public:
     MinMax(SEXP x, bool is_summary_ = false) :
       Base(x),
@@ -79,7 +26,43 @@ namespace dplyr {
 
     STORAGE process_chunk(const SlicingIndex& indices) {
       if (is_summary) return data_ptr[ indices.group() ];
-      return ChunkProcessor(data_ptr, indices).process();
+
+      const int n = indices.size();
+      STORAGE res = Inf;
+
+      bool first = true;
+      for (int i = 0; i < n; ++i) {
+        STORAGE current = data_ptr[indices[i]];
+
+        if (Rcpp::Vector<RTYPE>::is_na(current)) {
+          if (NA_RM)
+            continue;
+          else
+            return current;
+        }
+        else {
+          // Need a flag here, because is_better() cannot compare with Inf for RTYPE == INTSXP.
+          // Hoping that compiler is able to split this in two loops, or that branch prediction
+          // will make this check cheap.
+          if (first) {
+            res = current;
+            first = false;
+          }
+          else {
+            if (is_better(current, res))
+              res = current;
+          }
+        }
+      }
+
+      return res;
+    }
+
+    inline static bool is_better(const STORAGE current, const STORAGE res) {
+      if (MINIMUM)
+        return internal::is_smaller<RTYPE>(current, res);
+      else
+        return internal::is_smaller<RTYPE>(res, current);
     }
 
   private: