--- a/inst/include/dplyr/Result/all.h
+++ b/inst/include/dplyr/Result/all.h
@@ -12,8 +12,7 @@
 #include <dplyr/Result/Sum.h>
 #include <dplyr/Result/Var.h>
 #include <dplyr/Result/Sd.h>
-#include <dplyr/Result/Min.h>
-#include <dplyr/Result/Max.h>
+#include <dplyr/Result/MinMax.h>
 
 #include <dplyr/Result/CallElementProxy.h>
 #include <dplyr/Result/DelayedProcessor.h>