… inst/include/dplyr/Result/Max.h
… inst/include/dplyr/Result/Min.h