Now handled by a single template class, with support for all corner cases (empty vector, NA values, ...). For integer input, `NA` is returned instead of `Inf`.

Fixes #2305.

@hadley: PTAL.