@hadley 2017-02-19
How hard would it be expose this directly as an R function? I'd be interested in a benchmark vs base R and the previous version. 

Maybe having four templated alternatives (missing yes/no + int/double) would be faster and lead to simple code (albeit with a small amount of duplicated code)