@krlmlr 2017-02-19
It's already there: `with_hybrid(max(a), a = 1:5)` vs. `without_hybrid((max)(a), a = 1:5)`. (Just fixed a few problems in master.)

## Current master

``` r
library(dplyr, warn.conflicts = FALSE)
N <- 1e7
x <- runif(N)
xx <- x
xx[N/2] <- NA

microbenchmark::microbenchmark(
  dplyr:::with_hybrid(max(x), x),
  dplyr:::without_hybrid((max)(x), x),
  dplyr:::with_hybrid(max(x, na.rm = TRUE), x),
  dplyr:::without_hybrid((max)(x, na.rm = TRUE), x),
  times = 20
)
#> Unit: milliseconds
#>                                               expr      min       lq
#>                     dplyr:::with_hybrid(max(x), x) 55.14655 55.65483
#>                dplyr:::without_hybrid((max)(x), x) 13.73558 14.04404
#>       dplyr:::with_hybrid(max(x, na.rm = TRUE), x) 55.05419 55.68407
#>  dplyr:::without_hybrid((max)(x, na.rm = TRUE), x) 13.48640 13.88385
#>      mean   median       uq      max neval cld
#>  57.78299 56.72489 59.26558 64.88293    20   b
#>  14.80859 14.16043 14.80196 23.42316    20  a 
#>  58.49981 56.25381 59.57602 73.41693    20   b
#>  14.32813 14.00952 14.37358 17.76009    20  a
```

## This PR

``` r
#> Unit: milliseconds
#>                                               expr      min       lq
#>                     dplyr:::with_hybrid(max(x), x) 60.99932 62.36870
#>                dplyr:::without_hybrid((max)(x), x) 13.60645 14.00775
#>       dplyr:::with_hybrid(max(x, na.rm = TRUE), x) 55.21039 55.84697
#>  dplyr:::without_hybrid((max)(x, na.rm = TRUE), x) 13.65285 14.29137
#>      mean   median       uq      max neval cld
#>  66.16124 64.89561 69.02802 75.90574    20   c
#>  15.79797 15.29058 17.63915 18.72142    20 a  
#>  58.31746 56.99891 58.81414 75.32671    20  b 
#>  15.89408 15.05983 16.87834 27.06653    20 a
```

The `na.rm` case has become a tad slower with this change (with `gcc -O2` on my laptop), overall our implementation is slower than base R's by a factor of ~3-4.

I combined the four classes into one to reduce maintenance effort (now and in the future). I'd argue that a minimal run time improvement isn't worth the extra maintenance effort, given that we're already sub-par. I could look at a profiler run if it's important, though.