@hadley 2017-02-19
If the hybrid is so much slower, why do we even have it?