@krlmlr 2017-02-19
It's faster in a grouped query, because we avoid the roundtrip to the R interpreter and the materialization of the groupwise column values.