@hadley 2017-02-20
Base R is pretty inconsistent:

``` r
typeof(median(FALSE))
#> [1] "logical"
typeof(median(c(FALSE, TRUE)))
#> [1] "double"

typeof(median(1L))
#> [1] "integer"
typeof(median(c(1L, 2L)))
#> [1] "double"
```

I think we should say that the hybrid median _always_ returns a double.