@krlmlr 2017-02-20
This PR only touches hybrid min and max.