@hadley 2017-02-20
Oops. The equivalent base R code is this:

``` r
typeof(min(1L, na.rm = TRUE))
#> [1] "integer"
typeof(min(NA_integer_, na.rm = TRUE))
#> Warning in min(NA_integer_, na.rm = TRUE): no non-missing arguments to min;
#> returning Inf
#> [1] "double"

typeof(min(TRUE, na.rm = TRUE))
#> [1] "integer"
typeof(min(NA, na.rm = TRUE))
#> Warning in min(NA, na.rm = TRUE): no non-missing arguments to min;
#> returning Inf
#> [1] "double"
```

Seems like it would be more consistent to always return a numeric.