@krlmlr 2017-02-20
Now always returning numeric. New times:

``` r
#> Unit: milliseconds
#>                                                 expr      min       lq
#>                       dplyr:::with_hybrid(max(x), x) 57.92209 58.49885
#>                  dplyr:::without_hybrid((max)(x), x) 13.72525 13.98664
#>                     dplyr:::with_hybrid(max(xx), xx) 29.64903 30.68764
#>                dplyr:::without_hybrid((max)(xx), xx) 13.82268 14.15346
#>       dplyr:::with_hybrid(max(xx, na.rm = TRUE), xx) 55.57600 55.79007
#>  dplyr:::without_hybrid((max)(xx, na.rm = TRUE), xx) 13.72621 14.23667
#>      mean   median       uq      max neval  cld
#>  60.72233 59.46146 62.08578 67.16394    20    d
#>  14.77884 14.19997 14.55969 22.61328    20 a   
#>  33.49701 32.23387 33.76267 60.47334    20  b  
#>  15.08331 14.86969 15.44266 18.15982    20 a   
#>  57.59341 56.32895 58.95452 63.45070    20   c 
#>  14.88210 14.83992 15.22170 17.00810    20 a
```